requirements.txt
PyYAML
torch
dacite
torchaudio
faster-whisper
soundfile